Added requirements file and organised folders
Added __init__.py file for later
