--- test2-1.txt
@@ -1,14 +1,0 @@
-mode: chaser
-nb_turns: 0
-width: 11
-heigth: 12
-initial_nb_condemned_cells: 5
-actual_nb_condemned_cells: 5
-level: 0
-fugitive_x_position: 6
-fugitive_y_position: 11
-6 5
-5 1
-7 8
-0 7
-8 11
