+++ test_files/test3-2.txt
@@ -1,0 +1,61 @@
+mode: chaser
+nb_turns: 0
+width: 11
+heigth: 12
+initial_nb_condemned_cells: 5
+actual_nb_condemned_cells: 52
+level: 2
+fugitive_x_position: 5
+fugitive_y_position: 5
+0 1
+0 2
+0 3
+0 4
+0 5
+0 6
+0 7
+0 8
+0 9
+0 10
+1 2
+7 7
+5 1
+6 2
+2 3
+10 6
+10 5
+10 4
+10 3
+10 2
+10 1
+10 7
+10 8
+10 9
+10 10
+5 11
+6 11
+7 10
+4 11
+3 10
+2 10
+4 2
+7 2
+8 3
+2 8
+2 7
+2 6
+3 8
+4 9
+2 5
+6 9
+7 8
+8 7
+8 6
+7 4
+7 5
+4 4
+5 7
+4 7
+5 3
+2 1
+3 0
